piccoli fix e impaginazione albo index

+++ app/Http/Controllers/AlboController.php
@@ -22,7 +22,8 @@ class AlboController extends Controller
             'barcode' => 'required | integer',
             'prezzo' => 'required | min:0 | max:100',
         ]);
-
+        
+        /* Memorizzazione copertina su disco */
         $copertina = $request->file('copertina');
         $estensione = $copertina->getClientOriginalExtension();
         Storage::disk('public')->put($copertina->getFilename().'.'.$estensione, File::get($copertina));
@@ -64,17 +65,23 @@ class AlboController extends Controller
         
 
         $copertina = $request->file('copertina');
-        $estensione = $copertina->getClientOriginalExtension();
-        Storage::disk('public')->put($copertina->getFilename().'.'.$estensione, File::get($copertina));
+        if ($copertina) {
+            $estensione = $copertina->getClientOriginalExtension();
+            Storage::disk('public')->put($copertina->getFilename().'.'.$estensione, File::get($copertina));
+        }
 
         $albo = Albo::find($id);
         $albo->num_pagine = $request->get('num_pagine');
         $albo->prezzo = $request->get('prezzo');
         $albo->barcode = $request->get('barcode');
-        $albo->mime = $copertina->getClientMimeType();
-        $albo->original_filename = $copertina->getClientOriginalName();
-        Storage::disk('public')->delete($albo->filename);
-        $albo->filename = $copertina->getFilename().'.'.$estensione;
+        
+        if ($copertina) {
+            $albo->mime = $copertina->getClientMimeType();
+            $albo->original_filename = $copertina->getClientOriginalName();
+            Storage::disk('public')->delete($albo->filename);
+            $albo->filename = $copertina->getFilename().'.'.$estensione;
+        }
+        
         $albo->save ();
         return redirect(route('albo.index'))->with('success', 'L\'albo è stato aggiornato.');
     }
